gh-538 - Updated v2 examples service and added new resources to the v2 OperationsService.

+++ core/operation/src/test/java/uk/gov/gchq/gaffer/operation/export/set/GetSetExportTest.java
@@ -52,7 +52,7 @@ public class GetSetExportTest extends OperationTest {
         // Then
         assertEquals("key", deserialisedOp.getKey());
         assertEquals("jobId", deserialisedOp.getJobId());
-        assertEquals(0, deserialisedOp.getStart());
+        assertEquals(0, (int) deserialisedOp.getStart());
         assertEquals(5, (int) deserialisedOp.getEnd());
     }
 
@@ -70,7 +70,7 @@ public class GetSetExportTest extends OperationTest {
         // Then
         assertEquals("key", operation.getKey());
         assertEquals("jobId", operation.getJobId());
-        assertEquals(0, operation.getStart());
+        assertEquals(0, (int) operation.getStart());
         assertEquals(5, (int) operation.getEnd());
     }
 }
